+++ dunamai/__main__.py
@@ -43,11 +43,24 @@ def parse_args(argv=None) -> argparse.Namespace:
             " see Version.from_*() docs for more info"
         ),
     )
+    from_sp.add_argument(
+        "--format",
+        help=(
+            "Custom output format. Available substitutions:"
+            " {base}, {epoch}, {pre_type}, {pre_number}, {post}, {dev}, {commit}, {dirty}"
+        )
+    )
 
     return parser.parse_args(argv)
 
 
-def from_vcs(vcs: Vcs, pattern: Optional[str], with_metadata: Optional[bool], with_dirty: bool) -> None:
+def from_vcs(
+        vcs: Vcs,
+        pattern: Optional[str],
+        with_metadata: Optional[bool],
+        with_dirty: bool,
+        format: Optional[str],
+    ) -> None:
     callbacks = {
         Vcs.Any: Version.from_any_vcs,
         Vcs.Git: Version.from_git,
@@ -59,10 +72,10 @@ def from_vcs(vcs: Vcs, pattern: Optional[str], with_metadata: Optional[bool], wi
         arguments.append(pattern)
 
     version = callbacks[vcs](*arguments)
-    print(version.serialize(with_metadata, with_dirty))
+    print(version.serialize(with_metadata, with_dirty, format))
 
 
 def main() -> None:
     args = parse_args()
     if args.command == "from":
-        from_vcs(Vcs(args.vcs), args.pattern, args.metadata, args.dirty)
+        from_vcs(Vcs(args.vcs), args.pattern, args.metadata, args.dirty, args.format)